模块化，牺牲少量性能提升可读性
