+++ front_end/src/components/visualization/ActivityCalendarAbstract/Setting.vue
